Add BiddingHistory component for displaying contract bidding history
- Introduced a new BiddingHistory component to present bidding history in a structured table format.
- Implemented loading states and fallback content for improved user experience when no bids are available.
- Integrated the BiddingHistory component into ContractDetails, enhancing the overall contract management interface.
- Refactored existing code to remove redundant bidding history logic from ContractDetails, promoting better organization and maintainability.
